backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Blockchain
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# Biometrics
deepface==0.0.79
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
tensorflow==2.14.0
tf-keras==2.14.0
pillow==10.1.0

# Cryptography
cryptography==41.0.7
argon2-cffi==23.1.0
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
pyotp==2.9.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# CORS
python-cors==1.0.0

# Utilities
python-dateutil==2.8.2
